Visual C++ updates

+++ trunk/src/y-mk/yocto/math/ode/ode-solver-data.cxx
@@ -18,7 +18,7 @@ namespace yocto
             lw_arrays<real_t, memory_type>( 3 ),
             hmin(0),
             eps( numeric<real_t>:: ftol ),
-            TINY( 1e-30 ),
+            TINY( REAL(1e-30) ),
             y(     next_array() ),
             dydx(  next_array() ),
             yscal( next_array() )
